job_application_agents/requirements.txt
pyautogen
openai
beautifulsoup4
requests
selenium
python-dotenv